Fix the package structure of the exceptions
