+++ Exeptions/src/sumator/DemoEx1.java
@@ -1,4 +1,4 @@
-package Ex1_StringSumator;
+package sumator;
 
 import java.util.Scanner;
